.github/scripts/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
PyGithub>=1.58.2
pytest>=8.0.0
